requirements.txt
spotipy
customtkinter
python-dotenv
pillow
